fix-staging-perms: set group name from Puppet
The /srv/mediawiki-staging and /srv/patches used to be owned by the
`wikidev` group which was changed in 2022 to use `deployment`.

The `fix-staging-perms` can be sudoed by non root to reset permissions
and ownership in case something screw it up.  It hardcoded the
`wikidev` which eventually broke 1.41.0-wmf.12 deployment today.

4021ea6ecf115 updated the group name, but it should really be maintained
by Puppet (and ultimately from the `deployment_group: deployment` Hiera
variable.

This makes Puppet to create a config file with a single
`deployment_group` variable and use it in the rest of the script.

Bug: T338205
Hosts: P:mediawiki::deployment::server
Change-Id: I9a6e3d52495b4a17ce8ed083d7c7b097edf20c63