scap3: stop defaulting deployment_group to 'wikidev'
In 2022 we switched the MediaWiki deployment files to be in the group
`deployment`. While deploying 1.41.0-wmf.12 we still had some code
referring to `wikidev` which caused some havoc.

The group to use is defined by Hiera `deployment_group`, this change
makes all classes to require the parameter to be set instead of having
the erroneous legacy `wikidev`.

Since deployment-prep on WMCS has a scap3 deployment but still uses
`wikidev`, set `deployment_group: wikidev` for that project.

Bug: T338205
Hosts: P:mediawiki::deployment::server
Change-Id: I38bd72e061d7a827738ebd73036f9c89a3775775