fix-staging-perms: set set-group-id on /srv/patches subdirs
The `/srv/patches` is shared among user and as such Puppet sets the
set-group-id flag on the directory:

    file { $patches_path:
        ensure => 'directory',
        owner  => 'mwdeploy',
        group  => $deployment_group,
        mode   => '2775',   # <------ 2??? = set-group-id
    }

That is honored when scap copies the patches for the new MediaWiki
branch, however the existing `/srv/patches/.git` lack the flag, most
probably cause it was created earlier and we need to be able to fix it.

Adjust fix-staging-perms to:
* Apply set-group-id to `/srv/patches`
* Apply it to any sub directory present (including /srv/patches/.git).

The git repository is local to the host and not crafted by Puppet, so
there is no resource to adjust in Puppet to make it set the
set-group-id.

Once this patch has been deployed, run the `fix-staging-perms` on the
deployment server which will set the set-group-id flag on the git
directories:

  deploy1002:~$ find /srv/patches -type d -not -perm /g+s
  /srv/patches/.git
  /srv/patches/.git/objects
  /srv/patches/.git/objects/pack
  /srv/patches/.git/objects/info
  /srv/patches/.git/branches
  /srv/patches/.git/refs
  /srv/patches/.git/refs/heads
  /srv/patches/.git/refs/tags
  /srv/patches/.git/hooks
  /srv/patches/.git/info
  /srv/patches/.git/logs
  /srv/patches/.git/logs/refs
  /srv/patches/.git/logs/refs/heads

Bug: T338205
Hosts: P:mediawiki::deployment::server
Change-Id: Iac712717679b9cdb85a50eacfe2898c4138ddd48