Update the maintain-views script to improve the table selection option
We have a number of views that use multiple source tables, but when the
`--table` option is provided to the maintain-views script it only works
on views that contain a single source. Any custom views of this type are
skipped.

This change updates the filtering logic so that it also checks to see
whether the table name is contained within a list of the source tables,
rather than only the case where it is an exact match for a single item.

Bug: T315426
Bug: T311588
Change-Id: I4583d238c490b7b5fd6dcef7ab152b40669e790e