ntp: do not restart the ntp service on conf change
It's probably a bad idea that we let the ntp service restart on every
configuration file change. Using a more recent case as an example,
bringing up a new dnsbox during a hardware refresh causes the ntp_peers
to change which then subsequently restarts the ntp service on all DNS
hosts, much quicker (whenever there is an automatic Puppet run) than
what we would like it to.

Let's remove the restart change but instead add a check so that we
remember to restart the service manually, spacing it out as we like.

Hosts: P:ntp
Change-Id: Ibb4854c140d38e675a701eaaefd7ef8c5ec2ed22