zuul: remove mode/umask from config git clone
Docker-pkg and Zuul configuration are stored in integration/config.git
which is cloned on the contint server at `/etc/zuul/wikimedia`.

For historical reason, the git repository was cloned with:

  mode  => '0775'
  umask => '002'

That is a way to ensure git creates files that are group writable which
is helpful when a local workspace is shared between users. That set of
parameters with those exact values is superseeded by `shared => true`
which sets git config `core.sharedRepository=group` and has the same
effect as the above (mode, umask).

I have considered switching to `shared => true` (T338277), then I have
realized the repository is owned by `zuul:zuul` and our deployment
script in integration/config runs all commands as the `zuul` user
already. That is allowed by a sudo rule for members of `contint-admins`
group.

Given all git operations are made as `zuul` and all files are owned by
`zuul`, there is no need to make it shareable and we can remove the
legacy `mode` and `umask` parameters.

After deployment, one might want to remove the group writable bit on
files even though I don't think it is necessary:

 sudo chown -R g-w /etc/zuul/wikimedia

Bug: T338277
Hosts: P:zuul::server
Change-Id: I43af53f2b16758edaac4c64b15329ba2025b2ee1