admin: reserve gerrit uid/gid
Switching over the Gerrit services requires rsync between the hosts. Due
to rsync inability to map uid/gid when the daemon has "use chroot =
yes", we have to hardcode the uid/gid on all hosts.

I have reserved gerrit uid/gid 925 for that purpose:
https://wikitech.wikimedia.org/w/index.php?title=UID&diff=prev&oldid=2083729

NOTE: the application currently uses `gerrit2` which was used for
historical reason and should be renamed to `gerrit`. Since updating the
uid/gid needs a full crawl, we might as well switch the user and group
names at the same time.

Bug: T338470
Change-Id: Ifcb68e36983101937f32c93d7d2b64fa4fe45e59