tshark: drop debconf::seen
debconf::seen uses debconf-communicate to set the debconf value, but
debconf-communicate throws an error if the value does not exist, which
means we would need to install the package first, which is awkward.
Simply drop the debconf:seen, it's only purpose it to avoid a debconf
prompt when running apt interactively (puppet and debdeploy don't show
debconf prompts), which is a very rare operation.

Hosts: basic
Change-Id: I96ccde96f98738e7f80bf7da7022a8503b84a153