profile::cache::kafka::certificate: fix client PKI config
After reading
https://github.com/confluentinc/librdkafka/wiki/Using-SSL-with-librdkafka#configure-librdkafka-clientvar
I realized that I was wrong in the previous config for PKI,
since librdkafka (used by varnishkafka) doesn't really use keystores
for client auth, but key + cert files. From
https://github.com/confluentinc/librdkafka/issues/2647
it seems that the key password is optional, so we should be
good with PKI to just generate the key + cert file and
reference them in the Varnishkafka config.

Bug: T337825
Change-Id: Ic547227cff522d6a278abb28edb7ee248072b9c1