Revert "Bump mediawiki_history_reduced version for aqs"
This reverts commit 01b79192d5b4cb95b51a956576a1cb88857e8968.

Reason for revert: For some reason, this latest snapshot of
mediawiki_history_reduced is not showing up in Druid.  I tried re-running the
indexing job and it doesn't seem to error but it doesn't result in any more of
a good state.  Reverting for now until we figure out what's up with the job.
Without reverting to a working state, wikistats and AQS will be returning "no
data" for all the Druid-based metrics.

Change-Id: I4e439a65c82e35fad80afc920b5f9703958b499d