P:toolforge::apt_pinning: remove sssd pinning
There are security updates for sssd available on buster (as
1.16.3-3.2+deb10u2). The sssd pinning doesn't cover all of the sssd
packages, most notably it's missing python3-sss, libsss-* and
libipa-hbac*. So Apt wants to uninstall sssd in order to install
security updates for the rest of the packages, which is definitely not
wanted here.

Instead of fixing the pinning we drop it entirely to apply security
updates automatically.

Bug: T290494
Change-Id: I2d9cebfe2158f82ac61d4f72c53429427c43e792