trafficserver: route proton requests via the API gateway
As part of the restbase deprecation process we are rerouting on a
service by service basis to the rest-gateway. The rest-gateway is more
or less passing requests through to the service itself with minimal
transformation. Proton's simple API, cache behaviour and low traffic
rate makes it a good candidate for the first service.

Bug: T324678
Change-Id: Ic2547b1c2b4830aa3610cd168e81a02039755d7e