gitlab: Add locking to backups
Change gitlab backup and restore scripts to respect lock files. The
backup and restore process has been taking longer, and is running more
frequently, so manual work is interfering and being interfered with by
the cron jobs.

This ensures that only one restore or backup can run at any one time on
a host. We use time-based expiry of lock files because we want to create
lock files using the cookbook to block things on other hosts, so
tracking the pid won't work.

Additionally, move the templated variables to a separate config file to
avoid templating executable scripts

Change-Id: I6772c5b9fc362625f140e5eac3861b98abe30cb8