ferm: Allow passing the port in a more structured way
So far the $port is untyped and expects Ferm syntax, it's currently receives
in the form of
- strings
- integers
- the ferm internal range specifier
- in a few cases also in form of the service name (as used in /etc/services)

Extend the existing Ferm::Port type to also allow an array of Stdlib::Port
to switch uses of profile::firewall to a tool-agnostic syntax.

Also add a new type for a port range.

Hosts: C:profile::firewall
Bug: T336497
Change-Id: Ia9bb24699b9ec3e9e61c55547c33c8c371d0f7a5