acme_chief: openstack: codfw1dev: refresh LDAP certificates
The pair of cloudservices2004-dev / cloudservices2005-dev is losing the
wikimedia.org address. As part of this network setup rework we are
introducing a new service address ns-recursor.openstack.<deploy>.wikimediacloud.org
which we can use for the acme_chief certificate.

The slapd daemon as configure by puppet listens in 0.0.0.0:636 by default, so
no problem in that sense.

While at it, drop the ldap-labtest entry which is no longer required.

Bug: T307357
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I21212ce373bf93ea1ba954c5b20bd920796f0a91