Allow MGMT ranges to make TFTP requests to install server
Currently for our beta ZTP setup for Juniper devices we are serving
the init script using TFTP from our install servers.

This fails right now as the local iptables rules on the box is
blocking the traffic.  This patch adds $MGMT_NETWORKS to the list
of sources that can connect with tftp.

Hosts: install1004.wikimedia.org
Bug: T336485
Change-Id: Ib3d92ae6376d778be24ee454578df7e466194c13