d-i: Fix retrieval of reuse-parts.sh for bookworm
In early d-i wget is provided by busybox, which doesn't support HTTPS
in standard builds. Until Bullseye busybox transparently fell down
to http if an https URL is passed, but starting with the version in
bookworm (1.35) it complains about the URL:

wget: not an http or ftp url: https://apt.wikimedia.org/autoinstall/scripts/reuse-parts.sh

So simply switch to an HTTP URL (which was always used in practice). Our only
use for wget in early d-i is fetching some config files and it's not unlikely
that trixie will switch to growlight as a partman replacement, which will make
the whole code obsolete anyway.

Bug: T339835
Change-Id: I7722d09813e054c68681e24b982b60dc99baa3f6