aptrepo: Rename repo sync config used for megacli and apply to bookworm as well
Confusingly the repo sync definition is called hwraid-stretch for all suites.
This is mostly historical, in the past we had different imports for different
distros. These days the repo only still contains megacli, the others tools
provided in older suites on hwraid.le-vert.net (e.g. lsiutil, aacraid-status)
are totally obsolete.

As such, rename it to hwraid-megacli and also add it for bookworm.

Bug: T339847
Change-Id: I4827e825fd3ecd08b6fc2be1ad7dae182917973e