Update sync definition for HP raid modules for bookworm
For distros up to Bullseye we had imported all RAID modules,
but with the rework on the RAID classes (which now use explicit
PCI ID lookups to determine which controllers to install/activate)
we can reliably determine which controllers are actually in
use, outside of megacli that's just ssacli and perccli (the latter
gets installed via the private apt repo).

As such, hpssa hpssacli hpssaducli hpacucli ssaducli no longer
get imported for Bookworm and later.

In addition hp-health gets installed in base::standard_packages
for all distros up to buster, HPE never provided it for later
distros, so also drop it from the import (and retroactively also
apply that to the config for bullseye).

Bug: T339847
Change-Id: I5a161368eeadb21800b7e948254ebff8f9da6eb8