cassandra: add initial support for PKI TLS certs to 4.x
Cassandra 4+ seems to support hot-reload of TLS certificates,
so it seems nice to start with this version to experiment with PKI.

This patch adds the support for PKI for a new cluster, not running
with TLS settings. In order to support this use case, we'll need
to create a truststore that contains the PKI root tls cert
and the one currently used by a given cluster, so that cassandra
daemons will be able to accept certs from both when validating
TLS conns from other cassandra nodes.

We can start with this initial support, apply it to a cluster
like ml-cache, and then add supports for other clusters if
we feel so.

Bug: T288470
Change-Id: Ic906a1d30f4a7a69b5b662a29bb042fe4386638a