Retire legacy "raid" fact
This has been replaced by the "raid_mgmt_tools" fact in our tooling. In contrast
to the legacy "raid" fact, the replacement operates solely on PCI IDs which allows
for a more reliable mapping of RAID controllers towards the respective userspace
tooling. The old approach based on heuristics on device names became more and more
brittle.

Bug: T313312
Change-Id: I33d67889cc7cef5f4d4a152811a3c2091cab51e9