cloudnfs: refactor configuration
Splits the nfs-mounts.yaml.erb, which was previously used both by Puppet
to load configuration, and by the nfs-exportd daemon as a config file,
to a new config file (modules/cloudnfs/data/projects.yaml) which can be
edited by humans, syntax-checked, etc. and a new define
(cloudnfs::volume_config) which generates the config file used by a few
different scripts in the repository.

It also replaces the mount_nfs_volume function (which was using the old
Ruby function API) with a Puppet-native function and otherwise cleans up
the NFS mount logic.

The 'volumes_served' config option seems unused and is removed.

Change-Id: I1a63821ad9a37d767b3565c5d19da7c95ecaf8f9