gerrit: backup /home on gerrit servers in Bacula
As Antoine has said there might be scripts in home directories
on Gerrit servers that should be saved.

So if that is the case, let's just add the home dirs to backups.

And/or scripts should be puppetized and be copied during migrations.

But to be pragmatic about the status quo.. first need to backup
somewhere.

Also to make this even work we have to add the gerrit role BACK
to the old machine.. but we would also have to do that for a manual
rsync.. so .. it's just unfortunate we haven't done that already during
the recent migration.

Bug: T336427
Change-Id: I465deb8889c6f096f1986e88ccbf54215db23b9b