beta: avoid erasing extensions when already present
c707c24768f made git::clone to set the remote origin, that causes
Git::Clone[beta-mediawiki-core] to be considered refresh which in turns
triggers a `rm -r php-master/extensions`.

The refresh apparently occurs on every Puppet run, which deletes the
extensions and enters in a race condition with scap deploying mediawiki
every few minutes.

Add a guard to only delete `php-master/extensions` when it is not a git
directory.

NOTE: skins have a similar issue but that is handled differently using
`git init`. The code should be made common between extensions and skins
but that requires a bit more thinking.

Bug: T340030
Hosts: deployment-deploy03.deployment-prep.eqiad1.wikimedia.cloud
Change-Id: Ie48d49f444ea5397db3a41c2efac704d78b7edc4