wmcs-cinder-backup-manager.py: only full backups of giant volumes
Our backup nodes run out of memory when trying to compute diffs for
terrabyte-scale volumes. I assume this is something wrong with the
diffing algorithm but for now let's try just doing full backups
of these giants and see if that keeps things happy.

If this patch works it could use a followup: either fixing the backup
algorithm or moving this arbitary 5000 into a config setting.

Bug: T339830
Change-Id: I8a4301150b8d1c48a4dda3044c06232703c6632a