hiera: set ms-be2068 to be an object expirer
This will enable the swift object expirer on this host for the codfw
cluster (similar has already been done in deployment-prep). Assuming
this goes well, will then do similar for an eqiad host.

We might want to refactor this to be more like how we select the
rclone host, but don't want to delay the KR.

Bug: T229584
Change-Id: I99d640ad8f31f91f7a08bcc366ebf8251a857e99