ferm: Allow passing sets to an srange or drange
Add an explicit argument for passing a set (the terminology comes
from nft, the equivalent in Ferm is a @def macro. Set seemed a little
more generic so it was picked sa the umbrella term).

This allows to pass these as explicit values rather than providing
them as a general purpose Ferm srange string. Individual hosts/networks
and sets can be combined in a single service.

Hosts: C:profile::firewall
Bug: T336497
Change-Id: Ia375e519cdcd53c8da983a0302105e98c8c4e4c1