Enable the PRESTO_EXPAND_DATA feature flag in Superset
When this feature is enabled, nested types in Presto will be expanded into
extra columns and/or arrays. This commit enables the flag uniformly on
both our production and staging instance.

An alternative method would be to use a hiera parameter to determine
whether it is enabled, but that might be necessary.

Bug: T340144
Change-Id: I40ad1bff84dd88da1e0accadc9953d7a34757493