wmcs: Configure sudo not to prompt for passwords
Users are often confused when they are prompted for their password when
running `sudo` in the Cloud VPS environment. Such prompts generally
happen either because there has been a configuration lookup failure
resulting in a fallback to default `sudo` behavior (display lecture
warning and prompt for user's system password) or because the user is
not actually allowed to perform the `sudo` operation. In either case we
really do not want users to be entering their Developer account
passwords in the Cloud VPS environment. Instead it would be nicer to
tell them that they cannot do what they attempted.

Use a combination of the `passwd_tries` and `lecture` Default_Entry
options to disable password prompting and the default "We trust you have
received..." preamble. With this config in place a user who has failed
a NOPASSWD check on the sudo operation in question will be given an
error message that looks something like:

  Sorry, user $USER is not allowed to execute '$CMD' as $TARGET on $HOST

Bug: T205463
Change-Id: Ieff371f29e93d57fca2274e4494734d031f60b01