P:netbox Redirect to idp on OIDC auth
The Social Auth implementation in Netbox does not
respect the LOGIN_URL setting in Django, meaning
that users will be presented with a username/password
prompt that doesn't work and they will need to know
to click "cas" on that page. Instead let Apache
redirect the user to the correct login page.

Change-Id: I1f99828e2349cf31771e24a7b45791980b40af61