cassandra: add support for shorter TLS cert expiry checks
With PKI and Cassandra 4 we will have certificates with a shorter
expiry time (a month usually) that will be automatically picked
up and reloaded. The current monitors alert when Cassandra runs
a TLS certificate with an expiry time of 30 or less days, so
we need to define a new alert for the PKI use case.

Bug: T288470
Change-Id: Ib1a02a50c6c677cc88da4c585eefc6edeb54656c