releases-jenkins: replace Apache 2.2 with 2.4 syntax for access control
The goal is to stop using mod_access_compat which is the only
reason the old syntax still works.

So to ban by user agent (env) we have to use this new construct
inside a <RequireAll> and with a negation (Require not).

We need the <RequireAll> because default is <RequireAny> and in
that a negation would not have an effect.

Bug: T338071
Change-Id: I645881d9105f697fd81db956fb028105f30a4bf4