contint: replace Apache 2.2 access control syntax for Jenkins proxy
The goal is to stop using mod_access_compat which is the only
reason the old syntax still works.

So to ban by user agent (env) we have to use this new construct
inside a <RequireAll> and with a negation (Require not).

We need the <RequireAll> because default is <RequireAny> and in
that a negation would not have an effect.

Bug: T338071
Change-Id: Ib2ea283bd58dca1d395da79afd9e026c41d2e701