Add a workaround for a kerberos issue affecting Presto version 0.281
This is required in order to fix an issue with Presto accessing HDFS
after the upgrade to 0.281. The issue is being tracked upstream at
https://github.com/prestodb/presto/issues/18474 and we may want to
remove this setting once is fixed.

The switch disables a mechanism that the upstream authors have
introduced in order to try to improve the performance of kerberized
Hive/HDFS.

Bug: T337335
Change-Id: Ibc409ced4f77b4332ed719a9019d6ccdb42a8f53