P:bird::anycast_healthchecker: allow binding to multiple services
Our current anycast setup assumes that a single systemd unit can/should
be binded to the anycast-healthchecker service. Recently in discussions
around Wikimedia DNS, we decided we wanted to bind two services instead,
and so this is a refactor of P:bird::anycast_healthchecker to be able to
do that. There might be more cases that require this, and if not, the
change is simple enough (converting the string of a service to an array
of services) that we can update all configurations that set
profile::bird::bind_anycast_service.

We start with Wikimedia DNS and this will require a coordinated rollout
to ensure a NOOP on the existing hosts that called the profile.

Hosts: auto
Change-Id: I6a0282943f8dbd82c194784318f16ab9d88389a6