cumin: Properly set connect_timeout
I removed all the occurrences of ConnectTimeout in ssh_config files used
by Cumin, because Clustershell (the Python library that Cumin uses to
ssh) always overwrites the value found in the ssh_config with its own
config value "connect_timeout". If "connect_timeout" is not passed as a
config option, Clustershell sets it by default at 10.

This patch is setting connect_timeout:10 for cumin and cuminunpriv hosts
(that is the effective value that has been used by Cumin until now, as
"ConnectTimeout 2" was ignored).

In cloudcumin hosts we need a higher value connect_timeout:20 because
connections through the cloud bastion host can be slow, until T340241 is
fixed.

Example of a command that fails with a timeout of 10, but works with a
timout of 20, when run in cloudcumin1001:

  cumin -b30 -p0 --trace 'O{project:tools}' 'id'

Bug: T323484
Hosts: cumin1001.eqiad.wmnet,cuminunpriv1001.eqiad.wmnet
Hosts: cloudcumin1001.eqiad.wmnet
Hosts: cloud-cumin-03.cloudinfra.eqiad1.wikimedia.cloud
Change-Id: Id7d12c7f9eb7a9e3465faa50b1924b0decc87aa1