pybal: Make check conform to the Nagios plugin API
A previous change refactored the icinga check so that it displays more
information. However the refactor didn't take account that we only need
to send CRITICAL once and instead sends multiple messages. This patch
updates the check so the first line is just generic and then the next
lines provide a consolidated line detailing issues.

Related-Id: I88e4ed8ca3e7d2ce2b53701b5d0961b8917f43c7

Bug: T322377
Change-Id: I2de32fa919c7183e3ee6772efa0bfe813ca63be7