urldownloader: Simplify wikimedia variable
Why:
It's already out of date, doesn't use current data to begin with (e.g.
lacks the e/f rows in eqiad) and copies the data from the network
module, making it error prone and easily left stale

What:
Ditch the current, already wrong and outdated structure and just use
$mw_appserver_networks, which should contain all the current entries
plus the newer ones.

In a previous incarnation of this patchset, we were adding the newer
Kubernetes clusters. This is now handled at the network module layers,
so we are removing even those in this patch

Change-Id: If52176159ba99e3b42042385cd5bcc91aeccde8c
Hosts: auto