urldownloader: Switch $towikimedia to $mw_appserver_networks_private
Why:
We want to allow some use cases, like .e.g. ExtensionDistributor
reaching out to public hosts like Gerrit and Gitlab. The situation today
is very different to when this rule was originally created. The public
networks in the original list no longer have the workloads (i.e. mw
hosts) that this rule was meant to protect, as most workloads are now in
the private networks and are exposed solely via the CDN. The few
exceptions there, are treated anyway as external things by design and are
exposed on the internet relying on their own configurations for
protection

What:
Switch $towikimedia to $mw_appserver_networks_private

Change-Id: If2b6eb5b7833d197b69dab0e22d177394815e5a2
Hosts: auto