P:systemd::timesyncd: automate generation of ntp_servers list
We are currently managing the ntp_servers array in P:systemd::timesyncd
manually via hiera, where the client list is referenced from
hieradata/::site/profile/systemd/timesyncd.yaml. DNS/NTP hosts refreshes
involve manually updating this file but this should be automated and
this commit does that by shifting the logic to the profile instead of
the per-site hieradata.

To determine the nearest core site that is used for this purpose, we
introduce site_nearest_core in common.yaml.

Hosts: P:systemd::timesyncd
Bug: T340479
Change-Id: Ida2b4cb39514f95462c13343927085a774bb6165