role::cache::text: add ores-legacy.wikimedia.org
The ML team created a k8s-based replica of ores.wikimedia.org,
and we'd like to expose it publicly to testing it end-to-end.

The idea is to eventually add a CNAME for ores.w.o to ores-legacy.w.o,
so that we'll be able to deprecate the ORES workers bare metal
infrastructure.

Bug: T330414
Change-Id: Ie509bfb2d81b0cc54b3871040d1079d0cca9d851