puppet::agent: set manage_puppet_ca_file false
This variable was originally introduced to assist with extending the
lifetime of the current puppet CA.  however it is currently getting in
the way of moving clients from the current puppet infrastructure to the
new infrastructure.

The ca file is automatically installed by puppet when a client is
registered so changing this parameter should be mostly a safe no-op.

Hosts: basic
Hosts: P:puppetmaster::frontend
Change-Id: I5a0c1994d8573bf04ff987669cb8b5d42d3e9a0d