conftool: Add more servers to jobrunner cluster
Why:
The jobrunner cluster has been operating at higher
capacity. Additionally, we are migrating more
wikis to Using Parsoid in MW, instead of RESTbase,
so taffic is expected to increase.

How:
Moving servers from the api_appserver cluster to the
jobrunners cluster. api_appserver cluster is largely idle,
removing 5 servers is roughly -8% in number of hosts in it,
+25% in jobrunner hosts.

* Mark mw148[2-6].eqiad.wmnet as inactives
* Reimage said servers
* Pool back as jobrunners (and not videoscalers)

Bug: T329366
Change-Id: I627b905bb681bf1f54b2e86c6ac9928557dc9914