contint: parameterize the docker lvm disk size
The Jenkins agents have a 60GB ephemeral disk managed by LVM and
partitioned as:

* /var/lib/docker: 24G
* /srv: 36G (100%FREE)

The new instances have a 90GB disk and since the docker partition is
hardcoded to 24G they end up with a 60G /srv.

Add `size` parameter to `profile::ci::dockervolume` which we can use to
change the default 24G to 45G leaving 45G for /srv. On the new instance
we can then:

 profile::ci::dockervolume::size: '45G'

Bug: T340070
Change-Id: I861ff18c93c0ac144647b9af2a74eab19a4a3916