Add an apt mirror for the confluent-kafka 7.4 release
We would like to start working with confluent-kafka from their platform
7.4 release. This commit adds a new thirdparty repo called
thirdparty/confluent7 which targets the 7.4 release. It also filters the
repository to ensure that only the confluent-kakfa package is selected.
Unlike the confluent 4.0 release there is no version number in the
package name.

The archive signing key is different from that of the 4.0 release and
corresponds to http://packages.confluent.io/deb/7.4/archive.key

We have added it to the buster-wikimedia and bullseye-wikimedia
distributions. For the time being, we are only going to be testing the
use of the kafka clients, rather than any server components.

Bug: T329514
Bug: T300102
Bug: T302610
Change-Id: Id9adba32ae175c2b8185c8352cde078e86e137d1