Create spark3 local directory
jupyterhub fails to create a conda env on hosts running debian>11
due to missing spark3 local_dir /srv/spark-tmp.
The Local directory was being managed by the spark2 profile.
Since spark2 is no longer available for debian>11 we need to create the local directory.
We also need to remove the duplicate spark2 directory declaration.
This is also part of the spark2 deprecation.

Bug: T332765
Bug: T336036
Change-Id: Ic6790a9fa255dcb4bf24fa185b44d618b5b211f4