mw-on-k8s: Redirect 0.5% of all traffic to mw-on-k8s
Why:
As part of the mw-on-k8s migration, we want to keep directing more and
more traffic to the mw-on-k8s migration. group0 migration has been a
success, and we would like to move forward with greater diversity of
traffic. As such, we switch from group-based migration to
traffic-percentage-based.
We exclude wikis with known issues on k8s, such as votewiki, wikidata,
and commons.

What:
Change default ATS mw-on-k8s redirect probability from 0 to 0.005
Add matching rules for excluded wikis with a 0 redirect probability

Hosts: P:trafficserver::backend
Bug: T341078
Change-Id: I47697c0d5efb96413015067a4b4a425391831c3a