Configure spark3 defaults to use the new fetch protocol
We have had this compatability option in place so that we could submit
spark3 jobs and have them run by the spark2 shuffler service.

Now that we have enabled the spark3 shuffler, we may remove this
option.

See: https://spark.apache.org/docs/3.1.2/core-migration-guide.html for
more information on the use of spark.shuffle.useOldFetchProtocol

Bug: T332765
Change-Id: Ib4d30ca3772be41720dd8dac0dee58d71946c953