P:dns::recursor: automatically generate resolv.conf for DNS hosts
The resolv.conf data for the DNS hosts is currently manually generated
and set via per-host overrides for the DNS hosts. To reduce toil, this
list should be automatically generated. This commit aims to do that so
that future commissioning/decommissioning of the DNS hosts can
automatically pick up the required resolv.conf changes.

We build upon the existing work for the NTP peers as this (almost) uses
the same logic.

Bug: T340479
Hosts: cumin:P:dns::recursor
Change-Id: Ia5bf77c134f20666a2f7c53105ef8b890221f534