deployment_server: add REPL for mw-debug
We want members of the deployment group to be able to get an interactive
REPL shell inside the kubernetes environment.

To this end, create a script that can launch this repl when called with
"sudo", which launches either shell.php or eval.php similarly to what
deployers can do now using mwscript.

Given we're not really giving people the ability to execute scripts it
shouldn't be an issue do do this; at the same time, this is not a long-term
solution to the problem.

Bug: T341197
Change-Id: I473787a77c15e4d80cf2a100bb789f09d9f64f79