P:monitoring: improve check_service_restart.py
The check_service_restart.py is used in two places: in the NTP and the
Wikidough profiles. The purpose of this is to remind us to restart both
of these services in case the configuration file was changed, since we
do not let Puppet handle the restarts but we do so manually.

During the recent DNS host reimaging, it was observed that the NTP alert
was raised immediately. This happened because the script was making a
wrong assumption in just comparing the time delta between the file mtime
and the last service restart time. It is possible that a service was
restarted in the past but the file was only recently changed, so we
should not just compare that difference but should do so relative to the
current time (when the script is called).

This commit fixes that issues by checking if the current time and the
configuration file mtime exceeds the limit set by the critical argument
(defaults to 4 hours). If so, it raises a CRITICAL alert, otherwise it's
just a WARNING.

Hosts: dns1004.wikimedia.org, doh1001.wikimedia.org
Change-Id: Ie57f62eec7d637939b1a1fd5bcf5ac4f1bcc65a1