mw-on-k8s: Redirect 1% of all traffic to mw-on-k8s
Why:
As part of the mw-on-k8s migration, we want to keep directing more and
more traffic to the mw-on-k8s migration. group0 migration has been a
success, and we would like to move forward with greater diversity of
traffic. As such, we switch from group-based migration to
traffic-percentage-based.
We exclude wikis with known issues on k8s, such as votewiki, wikidata,
and commons.

What:
Change default ATS mw-on-k8s redirect probability from 0.005 to 0.01

Hosts: P:trafficserver::backend
Bug: T341463
Change-Id: I74bd60a1f9c2a629f639c38f7f21dfffc890f20a