hiera: add silent-drop directives for http frontend
Using different frontends and sticktables we should be able to split
the rate-limiting logic between tls and http frontend. For clarity,
ACLs keeps the same name across the two frontends.
No need to use different sticky counter names, as they are used in
different frontends and on different sticky tables.
The tls_terminator template has been updated to actually use the new
variables defined for frontend http and tls and the checks for the
tls/http keys existence has been moved to the template from the
manifests for better readability.
The concurrency limit for http frontend has been set to a fifth of the
frontend's maxconn on two canary hosts (both from text cluster, one
for esams and one for eqsin). Variables to track and debug behavior
has been set as for the TLS counterpart.

Bug: T340983
Change-Id: I8c00fa73c40a3d9375794f9cf3ccb5fea14e220f