Add a second karapace VM
So far, we have been using a single karapace VM to serve as a schema
registry service for both the production and staging deployments of
datahub.

This has now started to present problems of incompatible schemas being
served. Whilst in the long run it would be preferable to make use of the
new feature of an internal schema registry (within the GMS component),
for the time being we simply wish to separate the two karapace servers.

A follow-up commit will address the fact that the production karapace
instance uses the kafka-test cluster.

Bug: T329514
Bug: T341464
Change-Id: I4df3a6eb248f895191935d7b34a973f431adb284