webperf: Set XHGUI_PDO_INITSCHEMA=false to avoid 'CREATE TABLE' fatal
We currently run XHGui version 0.14. I'm attempting to upgrade
to 0.21.3 (latest) in Beta Cluster, but this fails due to a new
`CREATE TABLE IF NOT EXISTS` query that runs on every request to
the app, meant to lazily create/upgrade the schema.

Tracing back the versions, this first fails in 0.16.0, which moved
the schema lazy create feature in XHGui from its "save" class to
its general db "access" class, and thus runs on read requests as
well. This breaks in our deployment where we prefer to separate
admin from general read-write rights in MySQL grants.

I've patched this upstream at
https://github.com/perftools/xhgui/pull/494.

This patch sets the env variable to turn this feature off, thus
hopefully unlocking the XHGui upgrade in Beta and prod.

Change-Id: I652571783b9ef74e023757cfb4e8ab11acec6f98