Configure the test datahub jobs to use the staging schema registry
We have previoulsly been sharing a schema registry between the staging
and production jobs of datahub. This has started to cause an issue
regarding version upgrades, so we have now deployed a second instance of
karapace, backed by the kafka-test cluster, for datahub in staging.

This commit configures the airflow jobs in the analytics-test instance
to use this secondary schema registry when publishing their lineage.

Bug: T329514
Bug: T333004
Change-Id: I9b9d89f1e3c7c7cce2cb4face70c9806c9550b7a