replica_cnf_api: refactor to use multiple backends
This is a first step to get the envvars backend in there.

This changes slightly the return value for the delete operation, as
instead of just returning the deleted path, it returns the results for
each of the backends that tried to delete it.

This also adds the functional tests to the ci, and modifies them a bit
to be able to run both locally and on ci.

Bug: T265691
Signed-off-by: David Caro <dcaro@wikimedia.org>
Change-Id: I669901682d543be9e4b85a0517d978ada51388b9