burrow: use start-latest=true for the kafka consumer
From the upstream docs:
"""
If true, start consuming the offsets topic from the last (latest) offset.
Otherwise, start from the earliest offset.
"""
We have experienced out of memory errors when dealing with consumer
groups with too many offset committed, and starting from the earliest
seems to prevent the resolution of the issue (since the kafka consumer
will always try to pull data from the start).

For Burrow it should be fine to start from latest when the systemd
unit reloads.

Bug: T341551
Change-Id: I91b6159b56e96db5e4ca583d40d6d09ba3c72d28