test-cookbook: Add --no-sal-logging
Why:
In some cases testing cookbooks might require to run the cookbook
multiple times for real (not in dry-run mode) against a test environment
that is not affecting production in any way. In those cases it would be
useful to avoid logging to SAL to avoid the unnecessary IRC/SAL spam.

What:
Add a --no-sal-logging CLI argument that when set changes the spicerack
configuration file used to not log to SAL while printing a warning log
message to remind the user that SAL is disabled.

Change-Id: Idec906d87481ca742737823cfe99e3a7613e247d
Hosts: auto