confd: allow running multiple instances
We want to be able to run multiple instances of confd in parallel.

To this end:
* Make the confd class a simple class to install the basic components
* introduce a new confd::instance define that will actually run a confd instance
* Also introduce the class confd::default_instance that will replace the one previously
  declared in confd
* make the confd::file define require the default instance if no alternative instance
  is defined.

We kept most of our previous work untouched in this patch; in particular
I evaluated that the risk of naming clashes in the directory for bad outputs
is minimal. Same consideration was made for promehteus: given the risk of
naming being minimal, we didn't feel the need to change how prometheus
reporting works now.

Hosts: deploy1002.eqiad.wmnet
Hosts: mw1414.eqiad.wmnet
Bug: T341669
Change-Id: I805078c759f6bd92896a172bcc9d32cad9b44ec5