tox: drop the minor version requierment on admin checks
The admin and adminschema checks where set to use python3.7 explicitly
likely because it uses f-strings or something similar.  This patch drops
the requirement to use 3.7 explicitly as the majority of users should
no be on at least buster. This will enable users using python > 3.7 to
also run theses scripts

Change-Id: I3f9e5d27fac74483782747ee4a8c70611703ee1b