durum: bind anycast-healthchecker.service to nginx.service
Similar to the work done on the DNS and Wikimedia DNS hosts, add
bindings for anycast-hc on the durum hosts, binding it to the nginx
service. This helps ensure that if nginx is not running that the host is
depooled per the binding chain nginx -> anycast-hc -> bird. This is also
true for restarts and is the desired behaviour, given that there is no
durum service if nginx is not running.

Hosts: P:durum
Change-Id: I423544def7c52d4d0f17a952a72132c9ad14a232