Enable local caching for presto on the test cluster
These options for the presto coordinator and worker node should enable a
local instance of alluxio to be launched by each worker node, which
operates as a local query cache.

It is not a fully distributed alluxio caching layer, but is local to
each worker.

This change enables local caching on the test cluster only, for which we
have only one worker node.

Bug: T266641
Change-Id: I2f46de88ac811ffbf0a4c12a8763761b4344d0b8