Do not attempt to use hdparm on nvme drives for cephosd servers
The cephosd servers contain an NVMe drive that is used for caching
write access to hard drives and storing the WAL devices. Puppet was
previously configured to disable the write cache to the hard drives, but
inadvertently included the NVMe drive.

Attempting run hdparm on the NVMe drive exist with an error, so this
change causes us to diable the write cache on devices that match 'sd*'
and therefore excludes 'nvme*'

Bug: T330151
Change-Id: I196e00812b6e41ad8382a8d8400e63f6465a9a03